Refactor: Rename Room DB files and add FolderEntity
- Renamed `ReceiptDb.kt` to `RoomDb.kt` and `ReceiptDbModule.kt` to `RoomDbModule.kt` for better clarity.
- Moved `ReceiptEntity.kt`, `ReceiptDao.kt`, `ReceiptDbRepository.kt`, and `ReceiptAdapter.kt` to a new `receipt` subpackage within `room`.
- Added `FolderEntity` to `ReceiptEntity.kt` to represent folders for organizing receipts.
- Updated `ReceiptEntity` to include `folderId` to link receipts to folders.
- Renamed `ReceiptAppDatabase` to `ReceiptSplitterDatabase`.
