--- app/src/main/java/com/iliatokarev/receipt_splitter_app/receipt/data/room/ReceiptDb.kt
@@ -1,12 +1,0 @@
-package com.iliatokarev.receipt_splitter_app.receipt.data.room
-
-import androidx.room.Database
-import androidx.room.RoomDatabase
-
-@Database(
-    entities = [ReceiptEntity::class, OrderEntity::class],
-    version = 1
-)
-abstract class ReceiptAppDatabase : RoomDatabase() {
-    abstract fun receiptDao(): ReceiptDao
-}
